add category manager hook
